Оптимизация памяти и поддержка обучения на Paperspace
- Исправлена ошибка нехватки памяти при создании последовательностей (float32 вместо float64)
- Уменьшено использование памяти в 2 раза (с ~30 ГБ до ~15 ГБ для больших датасетов)
- Создан скрипт upload_to_paperspace.py для упаковки и отправки данных на Paperspace
- Добавлен requirements_linux.txt для установки зависимостей без MetaTrader5
- Добавлена документация по настройке обучения на Paperspace
- Обновлен CHANGELOG.md

+++ models/data_loader.py
@@ -145,7 +145,9 @@ class SequenceGenerator:
             sequences.append(sequence)
             targets.append(target)
         
-        return np.array(sequences), np.array(targets)
+        # Используем float32 вместо float64 для экономии памяти (в 2 раза меньше)
+        # Это критично для больших датасетов (155k+ образцов с 430 фичами)
+        return np.array(sequences, dtype=np.float32), np.array(targets, dtype=np.int64)
     
     def fit_scaler(self, df: pd.DataFrame):
         """Обучает scaler на данных"""
